separate out diloco configs (#1516)
