+++ torchtitan/components/ft/config/__init__.py
@@ -1,0 +1,13 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from torchtitan.components.ft.config.job_config import FaultTolerance, JobConfig
+
+
+__all__ = [
+    "FaultTolerance",
+    "JobConfig",
+]
