+++ torchtitan/models/__init__.py
@@ -9,6 +9,7 @@
 # will be called.
 
 import torchtitan.models.llama3  # noqa: F401
+import torchtitan.models.llama3_ft  # noqa: F401
 
 import importlib, warnings
